refactor: change from appdir -> pages
