+++ src/pages/_document.tsx
@@ -1,0 +1,15 @@
+import { Html, Head, Main, NextScript } from 'next/document'
+
+const Document = (): JSX.Element => {
+  return (
+    <Html className="scroll-smooth" lang="en">
+      <Head />
+      <body className="overflow-x-hidden antialiased">
+        <Main />
+        <NextScript />
+      </body>
+    </Html>
+  )
+}
+
+export default Document
